pkp/pkp-lib#8409 moved proxy classes into PKP namespace
