+++ classes/core/PKPContainer.php
@@ -27,7 +27,7 @@ use Illuminate\Queue\Failed\DatabaseFailedJobProvider;
 use Illuminate\Support\Facades\Facade;
 use PKP\config\Config;
 use PKP\i18n\LocaleServiceProvider;
-use PKP\Support\ProxyParser;
+use PKP\proxy\ProxyParser;
 
 use Throwable;
